+++ packages/dart_frog_cli/lib/src/commands/dev/dev.dart
@@ -5,8 +5,6 @@ import 'package:dart_frog_cli/src/command.dart';
 import 'package:dart_frog_cli/src/commands/commands.dart';
 import 'package:dart_frog_cli/src/commands/dev/templates/dart_frog_dev_server_bundle.dart';
 import 'package:dart_frog_cli/src/dev_server_runner/dev_server_runner.dart';
-import 'package:dart_frog_cli/src/runtime_compatibility.dart'
-    as runtime_compatibility;
 import 'package:mason/mason.dart';
 
 /// {@template dev_command}
@@ -17,14 +15,10 @@ class DevCommand extends DartFrogCommand {
   DevCommand({
     super.logger,
     GeneratorBuilder? generator,
-    DevServerRunnerBuilder? devServerRunnerBuilder,
-    runtime_compatibility.RuntimeCompatibilityCallback?
-        ensureRuntimeCompatibility,
-  })  : _ensureRuntimeCompatibility = ensureRuntimeCompatibility ??
-            runtime_compatibility.ensureRuntimeCompatibility,
-        _generator = generator ?? MasonGenerator.fromBundle,
-        _devServerRunnerBuilder =
-            devServerRunnerBuilder ?? DevServerRunner.new {
+    DevServerRunnerConstructor? devServerRunnerConstructor,
+  })  : _generator = generator ?? MasonGenerator.fromBundle,
+        _devServerRunnerConstructor =
+            devServerRunnerConstructor ?? DevServerRunner.new {
     argParser
       ..addOption(
         'port',
@@ -49,9 +43,7 @@ class DevCommand extends DartFrogCommand {
   static const _defaultDartVmServicePort = '8181';
 
   final GeneratorBuilder _generator;
-  final DevServerRunnerBuilder _devServerRunnerBuilder;
-  final runtime_compatibility.RuntimeCompatibilityCallback
-      _ensureRuntimeCompatibility;
+  final DevServerRunnerConstructor _devServerRunnerConstructor;
 
   @override
   final String description = 'Run a local development server.';
@@ -109,8 +101,6 @@ class DevCommand extends DartFrogCommand {
 
   @override
   Future<int> run() async {
-    _ensureRuntimeCompatibility(cwd);
-
     final port = io.Platform.environment['PORT'] ?? results['port'] as String;
 
     final dartVmServicePort = (results['dart-vm-service-port'] as String?) ??
@@ -130,7 +120,7 @@ class DevCommand extends DartFrogCommand {
       }
     }
 
-    _devServerRunner = _devServerRunnerBuilder(
+    _devServerRunner = _devServerRunnerConstructor(
       devServerBundleGenerator: generator,
       logger: logger,
       workingDirectory: cwd,